Comparative-Data-Analysis-and-Visualization-of-Employee-Data---Deep-Dive-With-pd-np-matplot-and-more.md
This project aims to perform a comprehensive comparative analysis of two employee datasets (df and df1). The analysis involves checking for similarities and differences between the two datasets, visualizing these differences using a variety of advanced plotting techniques, and providing a clear understanding of how the data varies across multiple dimensions. This project demonstrates the use of Python libraries such as Pandas, Matplotlib, Seaborn, and Plotly to explore, compare, and visualize data.